backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Pydantic
pydantic==2.9.2
pydantic-settings==2.5.2

# LangChain
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-google-genai==2.0.4
langchain-google-community==2.0.1
langgraph==0.2.45

# Vector Store
pgvector==0.3.5
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# PDF Processing
unstructured[pdf]==0.15.13
pytesseract==0.3.13
pdf2image==1.17.0
pillow==10.4.0

# Web Scraping
beautifulsoup4==4.12.3
httpx==0.27.2

# Search
tavily-python==0.5.0

# Utilities
python-dotenv==1.0.1
